\name{anon-package}
\alias{anon-package}
\docType{package}
\title{
\packageTitle{anon}
}
\description{
\packageDescription{anon}
}
\details{
(See help for individual functions.)
}
\author{
\packageAuthor{anon}

Maintainer: \packageMaintainer{anon}
}
\keyword{ package }
\keyword{ attribute }
