\name{neatoCurry-package}

\alias{neatoCurry-package}
\alias{neatoCurry}

\docType{package}

\title{ \packageTitle{neatoCurry} }

\description{ \packageDescription{neatoCurry} }

\details{

    \packageDESCRIPTION{neatoCurry}
    \packageIndices{neatoCurry}

}

\author{

    \packageAuthor{neatoCurry}

    Maintainer: \packageMaintainer{neatoCurry}
}

\keyword{ package }
