\name{zeroone-package}
\alias{zeroone-package}
\alias{zeroone}
\alias{O}
\alias{L}
\alias{OL}
\alias{LO}
\alias{LL}
\alias{LOL}
\docType{package}
\title{
\packageTitle{zeroone}
}
\description{
\packageDescription{zeroone}
}
\details{

\packageDESCRIPTION{zeroone}
\packageIndices{zeroone}
This package defines several integer vectors:
\code{L},
\code{O},
\code{LL},
\code{LO},
\code{OL},
and
\code{LOL}; defined as
\code{1L},
\code{0L},
\code{c(-1L, 1L)}, 
\code{1:0},
\code{0:1},
and
\code{-1:1},
}
\author{
\packageAuthor{zeroone}

Maintainer: \packageMaintainer{zeroone}
}
\references{
}
\keyword{ package }
\examples{
list(L=L, O=O, LL=LL, LO=LO, OL=OL, LOL=LOL)
}
